Speed improvements to LIST and windows console log.
- Fixed: List element values could only store 32 bits number, instead of 64 bits ones (like TAG).
- Improved: Windows console window log speed and removed rubberband effect when too much text was appended in a short amount of time.

Fixed item HP tooltip not updating when SelfRepair property kicks in.
Removed duplicate static GetIdentifierString function definition.

+++ src/sphere/ntservice.cpp
@@ -454,7 +454,7 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance, LPSTR lpCmdLine
 
 	TCHAR	*argv[32];
 	argv[0] = nullptr;
-	int argc = (int)Str_ParseCmds(lpCmdLine, &argv[1], CountOf(argv)-1, " \t") + 1;
+	int argc = Str_ParseCmds(lpCmdLine, &argv[1], CountOf(argv)-1, " \t") + 1;
 
 	// We need to find out what the server name is and the log files folder... look it up in the .ini file
     g_Cfg.LoadIni(false);
